Use soft deletes to maintain doc history (#29549)
Today we can use the soft-deletes feature from Lucene to maintain a
history of a document. This change simply replaces hard-deletes by
soft-deletes in Engine.

Besides marking a document as deleted, we also index a tombstone
associated with that delete operation. Storing delete tombstones allows
us to have a history of sequence-based operations which can serve in
recovery or rollback.

Relates #29530

+++ server/src/main/java/org/elasticsearch/index/engine/InternalEngine.java
@@ -63,6 +63,7 @@ import org.elasticsearch.index.IndexSettings;
 import org.elasticsearch.index.VersionType;
 import org.elasticsearch.index.mapper.IdFieldMapper;
 import org.elasticsearch.index.mapper.ParseContext;
+import org.elasticsearch.index.mapper.ParsedDocument;
 import org.elasticsearch.index.merge.MergeStats;
 import org.elasticsearch.index.merge.OnGoingMerge;
 import org.elasticsearch.index.seqno.LocalCheckpointTracker;
@@ -1220,7 +1221,17 @@ public class InternalEngine extends Engine {
             if (plan.currentlyDeleted == false) {
                 // any exception that comes from this is a either an ACE or a fatal exception there
                 // can't be any document failures  coming from this
-                indexWriter.deleteDocuments(delete.uid());
+                if (softDeleteEnabled) {
+                    final ParsedDocument tombstone = engineConfig.getTombstoneDocSupplier().newTombstoneDoc(delete.type(), delete.id());
+                    assert tombstone.docs().size() == 1 : "Tombstone doc should have single doc [" + tombstone + "]";
+                    tombstone.updateSeqID(plan.seqNoOfDeletion, delete.primaryTerm());
+                    tombstone.version().setLongValue(plan.versionOfDeletion);
+                    final ParseContext.Document doc = tombstone.docs().get(0);
+                    doc.add(softDeleteField);
+                    indexWriter.softUpdateDocument(delete.uid(), doc, softDeleteField);
+                } else {
+                    indexWriter.deleteDocuments(delete.uid());
+                }
                 numDocDeletes.inc();
             }
             versionMap.putUnderLock(delete.uid().bytes(),
